feat: [SIG-543]: refactor date time modal to separate component
