+++ frontend/src/components/CustomTimePicker/CustomTimePickerPopoverContent.tsx
@@ -1,6 +1,6 @@
 import './CustomTimePicker.styles.scss';
 
-import { Button, DatePicker } from 'antd';
+import { Button } from 'antd';
 import cx from 'classnames';
 import ROUTES from 'constants/routes';
 import { DateTimeRangeType } from 'container/TopNav/CustomDateTimeModal';
@@ -9,12 +9,10 @@ import {
 	Option,
 	RelativeDurationSuggestionOptions,
 } from 'container/TopNav/DateTimeSelectionV2/config';
-import dayjs, { Dayjs } from 'dayjs';
 import { Dispatch, SetStateAction, useMemo } from 'react';
-import { useSelector } from 'react-redux';
 import { useLocation } from 'react-router-dom';
-import { AppState } from 'store/reducers';
-import { GlobalReducer } from 'types/reducer/globalTime';
+
+import RangePickerModal from './RangePickerModal';
 
 interface CustomTimePickerPopoverContentProps {
 	options: any[];
@@ -40,35 +38,12 @@ function CustomTimePickerPopoverContent({
 	handleGoLive,
 	selectedTime,
 }: CustomTimePickerPopoverContentProps): JSX.Element {
-	const { RangePicker } = DatePicker;
 	const { pathname } = useLocation();
-
-	const { maxTime, minTime } = useSelector<AppState, GlobalReducer>(
-		(state) => state.globalTime,
-	);
 
 	const isLogsExplorerPage = useMemo(() => pathname === ROUTES.LOGS_EXPLORER, [
 		pathname,
 	]);
 
-	const disabledDate = (current: Dayjs): boolean => {
-		const currentDay = dayjs(current);
-		return currentDay.isAfter(dayjs());
-	};
-
-	const onPopoverClose = (visible: boolean): void => {
-		if (!visible) {
-			setCustomDTPickerVisible(false);
-		}
-		setIsOpen(visible);
-	};
-
-	const onModalOkHandler = (date_time: any): void => {
-		if (date_time?.[1]) {
-			onPopoverClose(false);
-		}
-		onCustomDateHandler(date_time, LexicalContext.CUSTOM_DATE_PICKER);
-	};
 	function getTimeChips(options: Option[]): JSX.Element {
 		return (
 			<div className="relative-date-time-section">
@@ -121,14 +96,11 @@ function CustomTimePickerPopoverContent({
 				)}
 			>
 				{selectedTime === 'custom' || customDateTimeVisible ? (
-					<RangePicker
-						disabledDate={disabledDate}
-						allowClear
-						onCalendarChange={onModalOkHandler}
-						// eslint-disable-next-line react/jsx-props-no-spreading
-						{...(selectedTime === 'custom' && {
-							defaultValue: [dayjs(minTime / 1000000), dayjs(maxTime / 1000000)],
-						})}
+					<RangePickerModal
+						setCustomDTPickerVisible={setCustomDTPickerVisible}
+						setIsOpen={setIsOpen}
+						onCustomDateHandler={onCustomDateHandler}
+						selectedTime={selectedTime}
 					/>
 				) : (
 					<div className="relative-times-container">
